Moved StataParser into new module pandas.io.stata
